Add code style to repo and fix it

See merge request fdroid/fdroidclient!1235

+++ app/src/main/java/org/fdroid/fdroid/installer/FileInstallerActivity.java
@@ -1,12 +1,17 @@
 package org.fdroid.fdroid.installer;
 
 import android.Manifest;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Bundle;
 import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.fragment.app.FragmentActivity;
 
 import org.apache.commons.io.FileUtils;
 import org.fdroid.fdroid.R;
@@ -17,12 +22,6 @@ import org.fdroid.fdroid.net.DownloaderService;
 
 import java.io.File;
 import java.io.IOException;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AlertDialog;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
-import androidx.fragment.app.FragmentActivity;
 
 public class FileInstallerActivity extends FragmentActivity {
 
@@ -105,22 +104,16 @@ public class FileInstallerActivity extends FragmentActivity {
         // pass the theme, it is not automatically applied due to activity's Theme.NoDisplay
         final AlertDialog.Builder builder = new AlertDialog.Builder(this, R.style.Theme_App);
         builder.setMessage(R.string.app_permission_storage)
-                .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        ActivityCompat.requestPermissions(activity,
-                                new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},
-                                MY_PERMISSIONS_REQUEST_STORAGE);
+                .setPositiveButton(R.string.ok, (dialog, id) -> ActivityCompat.requestPermissions(activity,
+                        new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                        MY_PERMISSIONS_REQUEST_STORAGE))
+                .setNegativeButton(R.string.cancel, (dialog, id) -> {
+                    if (act == 1) {
+                        installer.sendBroadcastInstall(canonicalUri, Installer.ACTION_INSTALL_INTERRUPTED);
+                    } else if (act == 2) {
+                        installer.sendBroadcastUninstall(Installer.ACTION_UNINSTALL_INTERRUPTED);
                     }
-                })
-                .setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        if (act == 1) {
-                            installer.sendBroadcastInstall(canonicalUri, Installer.ACTION_INSTALL_INTERRUPTED);
-                        } else if (act == 2) {
-                            installer.sendBroadcastUninstall(Installer.ACTION_UNINSTALL_INTERRUPTED);
-                        }
-                        finish();
-                    }
+                    finish();
                 })
                 .create().show();
     }
@@ -129,24 +122,22 @@ public class FileInstallerActivity extends FragmentActivity {
     public void onRequestPermissionsResult(int requestCode,
                                            @NonNull String[] permissions, @NonNull int[] grantResults) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        switch (requestCode) {
-            case MY_PERMISSIONS_REQUEST_STORAGE:
-                // If request is cancelled, the result arrays are empty.
-                if (grantResults.length > 0
-                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    if (act == 1) {
-                        installPackage(localApkUri, canonicalUri, apk);
-                    } else if (act == 2) {
-                        uninstallPackage(apk);
-                    }
-                } else {
-                    if (act == 1) {
-                        installer.sendBroadcastInstall(canonicalUri, Installer.ACTION_INSTALL_INTERRUPTED);
-                    } else if (act == 2) {
-                        installer.sendBroadcastUninstall(Installer.ACTION_UNINSTALL_INTERRUPTED);
-                    }
+        if (requestCode == MY_PERMISSIONS_REQUEST_STORAGE) { // If request is cancelled, the result arrays are empty.
+            if (grantResults.length > 0
+                    && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                if (act == 1) {
+                    installPackage(localApkUri, canonicalUri, apk);
+                } else if (act == 2) {
+                    uninstallPackage(apk);
                 }
-                finish();
+            } else {
+                if (act == 1) {
+                    installer.sendBroadcastInstall(canonicalUri, Installer.ACTION_INSTALL_INTERRUPTED);
+                } else if (act == 2) {
+                    installer.sendBroadcastUninstall(Installer.ACTION_UNINSTALL_INTERRUPTED);
+                }
+            }
+            finish();
         }
     }
 
@@ -163,7 +154,7 @@ public class FileInstallerActivity extends FragmentActivity {
         if (apk.isMediaInstalled(activity.getApplicationContext())) { // Copying worked
             Utils.debugLog(TAG, "Copying worked: " + localApkUri.getPath());
             if (!postInstall(canonicalUri, apk, path)) {
-                Toast.makeText(this, String.format(this.getString(R.string.app_installed_media), path.toString()),
+                Toast.makeText(this, String.format(this.getString(R.string.app_installed_media), path),
                         Toast.LENGTH_LONG).show();
                 installer.sendBroadcastInstall(canonicalUri, Installer.ACTION_INSTALL_COMPLETE);
             }
